@@ -1,11 +1,12 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import Amqp = require('./amqp');
-import Client = require('./client');
-
-declare class AmqpWs extends Amqp implements Client.Transport {
+import { Amqp } from './amqp';
+import { Client } from './Client';
+/**
+ * @class       module:azure-iothub.AmqpWs
+ * @classdesc   Constructs an {@linkcode Amqp} object that can be used in an application
+ *              to connect to IoT Hub instance, using the AMQP protocol over secure websockets.
+ *              This class overloads the constructor of the base {@link module:azure-iothub--amqp:Amqp} class from the AMQP transport, and inherits all methods from it.
+ */
+export declare class AmqpWs extends Amqp implements Client.Transport {
     constructor(config: Client.TransportConfigOptions);
+    protected _getConnectionUri(): string;
 }
-
-export = AmqpWs;
