@@ -1,18 +1,11 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 'use strict';
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
-Object.defineProperty(exports, "__esModule", { value: true });
-var amqp_1 = require("./amqp");
+
+import { Amqp } from './amqp';
+import { Client } from './Client';
+
 /**
  * @class       module:azure-iothub.AmqpWs
  * @classdesc   Constructs an {@linkcode Amqp} object that can be used in an application
@@ -25,15 +18,12 @@ var amqp_1 = require("./amqp");
 - `keyName` – (string) the name of a key that can be used to communicate with the IoT Hub instance
 - `sharedAccessSignature–` (string) the key associated with the key name.]*/
 /*Codes_SRS_NODE_IOTHUB_SERVICE_AMQP_WS_16_002: [`AmqpWs` should inherit from `Amqp`.]*/
-var AmqpWs = (function (_super) {
-    __extends(AmqpWs, _super);
-    function AmqpWs(config) {
-        return _super.call(this, config) || this;
-    }
-    AmqpWs.prototype._getConnectionUri = function () {
-        return 'wss://' + this._config.host + ':443/$iothub/websocket';
-    };
-    return AmqpWs;
-}(amqp_1.Amqp));
-exports.AmqpWs = AmqpWs;
-//# sourceMappingURL=amqp_ws.js.map
+export class AmqpWs extends Amqp implements Client.Transport {
+  constructor(config: Client.TransportConfigOptions) {
+    super(config);
+  }
+
+  protected _getConnectionUri(): string {
+    return 'wss://' + this._config.host + ':443/$iothub/websocket';
+  }
+}
