@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export import Client = require('./lib/client');
+export { Client } from './lib/client';
 export import ConnectionString = require('./lib/connection_string');
-export import Registry = require('./lib/registry');
-export import SharedAccessSignature = require('./lib/shared_access_signature');
-export import Amqp = require('./lib/amqp');
-export import AmqpWs = require('./lib/amqp_ws');
-export import DeviceMethodParams = require('./lib/device_method_params');
-export import JobClient = require('./lib/job_client');
-export import Device = require('./lib/device');
+export { Registry } from './lib/registry';
+export { SharedAccessSignature } from './lib/shared_access_signature';
+export { Amqp } from './lib/amqp';
+export { AmqpWs } from './lib/amqp_ws';
+export { DeviceMethodParams } from './lib/interfaces';
+export { JobClient } from './lib/job_client';
+export { Device } from './lib/device';
