@@ -1,6 +1,4 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-declare function translateError(message: string, amqpError: Error): Error;
-
-export = translateError;
+export interface AmqpTransportError extends Error {
+    amqpError?: Error;
+}
+export declare function translateError(message: string, amqpError: Error): AmqpTransportError;
