@@ -1,6 +1,2 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 import { ConnectionString } from 'azure-iot-common';
-
-export function parse(source: string): ConnectionString;
+export declare function parse(source: string): ConnectionString;
