@@ -1,10 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 'use strict';
-
-var _ = require('lodash');
-
+Object.defineProperty(exports, "__esModule", { value: true });
+var _ = require("lodash");
 /**
  * @deprecated
  * @class           module:azure-iothub.Device
@@ -38,47 +36,48 @@ var _ = require('lodash');
  * @prop {Object}   authentication  Contains the symmetric keys used to
  *                                  authenticate this device
  */
-
 /*Codes_SRS_NODE_SERVICE_DEVICE_16_001: [The constructor shall accept a `null` or `undefined` value as argument and create an empty `Device` object.]*/
-module.exports = function Device(jsonData) {
-  this.deviceId = null;
-  this.generationId = null;
-  this.etag = null;
-  this.connectionState = 'Disconnected';
-  this.status = 'Enabled';
-  this.statusReason = null;
-  this.connectionStateUpdatedTime = null;
-  this.statusUpdatedTime = null;
-  this.lastActivityTime = null;
-  this.cloudToDeviceMessageCount = 0;
-  this.authentication = {
-    symmetricKey: {  // <- this is the correct thing (camel-cased) 
-      primaryKey: null,
-      secondaryKey: null
-    },
-    x509Thumprint: {
-      primaryThumbprint: null,
-      secondaryThumbprint: null
+var Device = (function () {
+    function Device(jsonData) {
+        this.deviceId = null;
+        this.generationId = null;
+        this.etag = null;
+        this.connectionState = 'disconnected';
+        this.status = 'enabled';
+        this.statusReason = null;
+        this.connectionStateUpdatedTime = null;
+        this.statusUpdatedTime = null;
+        this.lastActivityTime = null;
+        this.cloudToDeviceMessageCount = '0';
+        this.authentication = {
+            symmetricKey: {
+                primaryKey: null,
+                secondaryKey: null
+            },
+            x509Thumbprint: {
+                primaryThumbprint: null,
+                secondaryThumbprint: null
+            }
+        };
+        /*Codes_SRS_NODE_SERVICE_DEVICE_16_002: [If the `deviceDescription` argument is provided as a string, it shall be parsed as JSON and the properties of the new `Device` object shall be populated with the values provided in the `deviceDescription` JSON string.]*/
+        /*Codes_SRS_NODE_SERVICE_DEVICE_16_003: [If the `deviceDescription` argument if provided as an object, the properties of the new `Device` object shall be populated with the values provided in the `deviceDescription` JSON string.]*/
+        if (jsonData) {
+            var userProps = (typeof jsonData === 'string') ? JSON.parse(jsonData) : jsonData;
+            if (!userProps.deviceId) {
+                /*Codes_SRS_NODE_SERVICE_DEVICE_16_004: [The constructor shall throw a `ReferenceError` if the `deviceDescription` argument doesn't contain a `deviceId` property.]*/
+                throw new ReferenceError('The \'deviceId\' property cannot be \'' + userProps.deviceId + '\'');
+            }
+            _.merge(this, userProps);
+        }
+        Object.defineProperty(this.authentication, 'SymmetricKey', {
+            enumerable: true,
+            get: function () {
+                /*Codes_SRS_NODE_SERVICE_DEVICE_16_005: [The `authentication.SymmetricKey` property shall return the content of the `authentication.symmetricKey` property (the latter being the valid property returned by the IoT hub in the device description).]*/
+                return this.symmetricKey;
+            }
+        });
     }
-  };
-
-  /*Codes_SRS_NODE_SERVICE_DEVICE_16_002: [If the `deviceDescription` argument is provided as a string, it shall be parsed as JSON and the properties of the new `Device` object shall be populated with the values provided in the `deviceDescription` JSON string.]*/
-  /*Codes_SRS_NODE_SERVICE_DEVICE_16_003: [If the `deviceDescription` argument if provided as an object, the properties of the new `Device` object shall be populated with the values provided in the `deviceDescription` JSON string.]*/
-  if(jsonData) {
-    var userProps = (typeof jsonData === 'string') ? JSON.parse(jsonData) : jsonData;
-    if (!userProps.deviceId) {
-      /*Codes_SRS_NODE_SERVICE_DEVICE_16_004: [The constructor shall throw a `ReferenceError` if the `deviceDescription` argument doesn't contain a `deviceId` property.]*/
-      throw new ReferenceError('The \'deviceId\' property cannot be \'' + userProps.deviceId + '\'');
-    }
-
-    _.merge(this, userProps);
-  }
-
-  Object.defineProperty(this.authentication, 'SymmetricKey', {
-    enumerable: true,
-    get: function () {
-      /*Codes_SRS_NODE_SERVICE_DEVICE_16_005: [The `authentication.SymmetricKey` property shall return the content of the `authentication.symmetricKey` property (the latter being the valid property returned by the IoT hub in the device description).]*/
-      return this.symmetricKey;
-    }
-  });
-};
+    return Device;
+}());
+exports.Device = Device;
+//# sourceMappingURL=device.js.map
