@@ -1,7 +1,5 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import { SharedAccessSignature } from 'azure-iot-common';
-
-export declare function create(host: string, policy: string, key: string, expiry: string | number): SharedAccessSignature;
-export declare function parse(source: string): SharedAccessSignature;
+import { SharedAccessSignature as Base } from 'azure-iot-common';
+export declare class SharedAccessSignature {
+    static create(host: string, policy: string, key: string, expiry: string | number): Base;
+    static parse(source: string): Base;
+}
