@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import { Registry, Device } from 'azure-iothub';
+import { Registry, Device } from '../../iothub';
 
 const connectionString = '[IoT Connection String]';
 
