@@ -5,7 +5,7 @@
 var sinon = require('sinon');
 var EventEmitter = require('events').EventEmitter;
 var util = require('util');
-var Amqp = require('../lib/amqp.js');
+var Amqp = require('../lib/amqp.js').Amqp;
 var assert = require('chai').assert;
 var SharedAccessSignature = require('azure-iot-common').SharedAccessSignature;
 var errors = require('azure-iot-common').errors;
