@@ -5,7 +5,7 @@
 
 var assert = require('chai').assert;
 var ArgumentError = require('azure-iot-common').errors.ArgumentError;
-var SharedAccessSignature = require('../lib/shared_access_signature.js');
+var SharedAccessSignature = require('../lib/shared_access_signature.js').SharedAccessSignature;
 
 var incompleteSignatures = {
   sr: 'SharedAccessSignature sig=signature&skn=keyname&se=expiry',
